+++ src/assets/styles/_variables.scss
@@ -1,10 +1,10 @@
-@import url('https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Lato&family=Montserrat&family=Playpen+Sans&family=Questrial&family=Roboto:wght@300&family=Shadows+Into+Light&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Amatic+SC&family=Indie+Flower&family=Poppins:wght@300&family=Silkscreen&family=Unica+One&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Lobster&display=swap');
 
 /* Fonts */
-$font-body: "Lato", sans-serif;
-$font-body-secondary: "Montserrat", sans-serif;
-$font-headers: "Shadows Into Light", sans-serif;
-$font-headers-secondary: "Bebas Neue", sans-serif;
+$font-body: "Poppins", sans-serif;
+$font-headers: "Indie Flower", sans-serif;
+$font-headers-secondary: "Lobster", sans-serif;
 
 /* Colour palette */
 $color-grey: #A9A9A9;
