+++ src/components/WeatherCard/WeatherCard.scss
@@ -1,0 +1,70 @@
+@use "src/assets/styles/variables" as *;
+
+.weathercard {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    padding: 1.5rem;
+    background: rgba(255, 255, 255, 0.435);
+    border-radius: 0.8rem;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    width: fit-content;
+
+    &__location {
+        font-size: 0.8rem;
+        margin-bottom: 0.5rem;
+        font-weight: 400;
+    }
+
+    &__local-time {
+        font-size: 0.6rem;
+        margin-bottom: 0.3rem;
+        font-style: italic;
+    }
+
+    &__local-date {
+        font-size: 0.6rem;
+        margin-bottom: 0.2rem;
+        font-style: italic;
+    }
+
+    &__conditions {
+        display: flex;
+        align-items: center;
+        gap: 0.8rem;
+    }
+
+    &__icon {
+        width: 64px;
+        height: 64px;
+    }
+
+    &__temp {
+        font-size: 0.7rem;
+    }
+
+    &__condition {
+        font-size: 0.7rem;
+    }
+
+    &__greeting {
+        font-family: $font-headers-secondary;
+        font-size: 1.2rem;
+        margin-bottom: 1rem;
+    }
+}
+
+.app-container {
+    display: flex;
+    justify-content: center;
+
+    &__loading {
+        margin-top: 7rem;
+        font-size: 0.7rem;
+    }
+}
